Crate rusoto_chime

The Amazon Chime API (application programming interface) is designed for administrators to use to perform key tasks, such as creating and managing Amazon Chime accounts and users. This guide provides detailed information about the Amazon Chime API, including operations, types, inputs and outputs, and error codes.

You can use an AWS SDK, the AWS Command Line Interface (AWS CLI), or the REST API to make API calls. We recommend using an AWS SDK or the AWS CLI. Each API operation includes links to information about using it with a language-specific AWS SDK or the AWS CLI.

Using an AWS SDK

You don't need to write code to calculate a signature for request authentication. The SDK clients authenticate your requests by using access keys that you provide. For more information about AWS SDKs, see the AWS Developer Center.

Using the AWS CLI

Use your access keys with the AWS CLI to make API calls. For information about setting up the AWS CLI, see Installing the AWS Command Line Interface in the AWS Command Line Interface User Guide. For a list of available Amazon Chime commands, see the Amazon Chime commands in the AWS CLI Command Reference.

Using REST API

If you use REST to make API calls, you must authenticate your request by providing a signature. Amazon Chime supports signature version 4. For more information, see Signature Version 4 Signing Process in the Amazon Web Services General Reference.

When making REST API calls, use the service name chime and REST endpoint https://service.chime.aws.amazon.com.

Administrative permissions are controlled using AWS Identity and Access Management (IAM). For more information, see Control Access to the Amazon Chime Console in the Amazon Chime Administration Guide.

If you're using the service, you're probably looking for ChimeClient and Chime.

Structs

Account
The Amazon Chime account details. An AWS account can have multiple Amazon Chime accounts.

AccountSettings
Settings related to the Amazon Chime account. This includes settings that start or stop remote control of shared screens, or start or stop the dial-out option in the Amazon Chime web application. For more information about these settings, see Use the Policies Page in the Amazon Chime Administration Guide.

BatchSuspendUserRequest

BatchSuspendUserResponse

BatchUnsuspendUserRequest

BatchUnsuspendUserResponse

BatchUpdateUserRequest

BatchUpdateUserResponse

ChimeClient
A client for the Amazon Chime API.

CreateAccountRequest

CreateAccountResponse

DeleteAccountRequest

DeleteAccountResponse

GetAccountRequest

GetAccountResponse

GetAccountSettingsRequest

GetAccountSettingsResponse

GetUserRequest

GetUserResponse

Invite
Invitation object returned after emailing users to invite them to join the Amazon Chime Team account.

InviteUsersRequest

InviteUsersResponse

ListAccountsRequest

ListAccountsResponse

ListUsersRequest

ListUsersResponse

LogoutUserRequest

LogoutUserResponse

ResetPersonalPINRequest

ResetPersonalPINResponse

UpdateAccountRequest

UpdateAccountResponse

UpdateAccountSettingsRequest

UpdateAccountSettingsResponse

UpdateUserRequest

UpdateUserRequestItem
The user ID and user fields to update, used with the BatchUpdateUser action.

UpdateUserResponse

User
The user on the Amazon Chime account.

UserError
The list of errors returned when errors are encountered during the BatchSuspendUser, BatchUnsuspendUser, or BatchUpdateUser actions. This includes user IDs, error codes, and error messages.

Enums

BatchSuspendUserError
Errors returned by BatchSuspendUser

BatchUnsuspendUserError
Errors returned by BatchUnsuspendUser

BatchUpdateUserError
Errors returned by BatchUpdateUser

CreateAccountError
Errors returned by CreateAccount

DeleteAccountError
Errors returned by DeleteAccount

GetAccountError
Errors returned by GetAccount

GetAccountSettingsError
Errors returned by GetAccountSettings

GetUserError
Errors returned by GetUser

InviteUsersError
Errors returned by InviteUsers

ListAccountsError
Errors returned by ListAccounts

ListUsersError
Errors returned by ListUsers

LogoutUserError
Errors returned by LogoutUser

ResetPersonalPINError
Errors returned by ResetPersonalPIN

UpdateAccountError
Errors returned by UpdateAccount

UpdateAccountSettingsError
Errors returned by UpdateAccountSettings

UpdateUserError
Errors returned by UpdateUser

Traits

Chime
Trait representing the capabilities of the Amazon Chime API. Amazon Chime clients implement this trait.