Trait rusoto_chime::Chime

pub trait Chime {
    fn batch_suspend_user(
        &self, 
        input: BatchSuspendUserRequest
    ) -> RusotoFuture<BatchSuspendUserResponse, BatchSuspendUserError>;

    fn batch_unsuspend_user(
        &self, 
        input: BatchUnsuspendUserRequest
    ) -> RusotoFuture<BatchUnsuspendUserResponse, BatchUnsuspendUserError>;

    fn batch_update_user(
        &self, 
        input: BatchUpdateUserRequest
    ) -> RusotoFuture<BatchUpdateUserResponse, BatchUpdateUserError>;

    fn create_account(
        &self, 
        input: CreateAccountRequest
    ) -> RusotoFuture<CreateAccountResponse, CreateAccountError>;

    fn delete_account(
        &self, 
        input: DeleteAccountRequest
    ) -> RusotoFuture<DeleteAccountResponse, DeleteAccountError>;

    fn get_account(
        &self, 
        input: GetAccountRequest
    ) -> RusotoFuture<GetAccountResponse, GetAccountError>;

    fn get_account_settings(
        &self, 
        input: GetAccountSettingsRequest
    ) -> RusotoFuture<GetAccountSettingsResponse, GetAccountSettingsError>;

    fn get_user(
        &self, 
        input: GetUserRequest
    ) -> RusotoFuture<GetUserResponse, GetUserError>;

    fn invite_users(
        &self, 
        input: InviteUsersRequest
    ) -> RusotoFuture<InviteUsersResponse, InviteUsersError>;

    fn list_accounts(
        &self, 
        input: ListAccountsRequest
    ) -> RusotoFuture<ListAccountsResponse, ListAccountsError>;

    fn list_users(
        &self, 
        input: ListUsersRequest
    ) -> RusotoFuture<ListUsersResponse, ListUsersError>;

    fn logout_user(
        &self, 
        input: LogoutUserRequest
    ) -> RusotoFuture<LogoutUserResponse, LogoutUserError>;

    fn reset_personal_pin(
        &self, 
        input: ResetPersonalPINRequest
    ) -> RusotoFuture<ResetPersonalPINResponse, ResetPersonalPINError>;

    fn update_account(
        &self, 
        input: UpdateAccountRequest
    ) -> RusotoFuture<UpdateAccountResponse, UpdateAccountError>;

    fn update_account_settings(
        &self, 
        input: UpdateAccountSettingsRequest
    ) -> RusotoFuture<UpdateAccountSettingsResponse, UpdateAccountSettingsError>;

    fn update_user(
        &self, 
        input: UpdateUserRequest
    ) -> RusotoFuture<UpdateUserResponse, UpdateUserError>;
}

Trait representing the capabilities of the Amazon Chime API. Amazon Chime clients implement this trait.

Required methods

fn batch_suspend_user(
&self,
input: BatchSuspendUserRequest
) -> RusotoFuture<BatchSuspendUserResponse, BatchSuspendUserError>

Suspends up to 50 users from a Team or EnterpriseLWA Amazon Chime account. For more information about different account types, see Managing Your Amazon Chime Accounts in the Amazon Chime Administration Guide.

Users suspended from a Team account are dissociated from the account, but they can continue to use Amazon Chime as free users. To remove the suspension from suspended Team account users, invite them to the Team account again. You can use the InviteUsers action to do so.

Users suspended from an EnterpriseLWA account are immediately signed out of Amazon Chime and are no longer able to sign in. To remove the suspension from suspended EnterpriseLWA account users, use the BatchUnsuspendUser action.

To sign out users without suspending them, use the LogoutUser action.

fn batch_unsuspend_user(
&self,
input: BatchUnsuspendUserRequest
) -> RusotoFuture<BatchUnsuspendUserResponse, BatchUnsuspendUserError>

Removes the suspension from up to 50 previously suspended users for the specified Amazon Chime EnterpriseLWA account. Only users on EnterpriseLWA accounts can be unsuspended using this action. For more information about different account types, see Managing Your Amazon Chime Accounts in the Amazon Chime Administration Guide.

Previously suspended users who are unsuspended using this action are returned to Registered status. Users who are not previously suspended are ignored.

fn batch_update_user(
&self,
input: BatchUpdateUserRequest
) -> RusotoFuture<BatchUpdateUserResponse, BatchUpdateUserError>

Updates user details within the UpdateUserRequestItem object for up to 20 users for the specified Amazon Chime account. Currently, only LicenseType updates are supported for this action.

fn create_account(
&self,
input: CreateAccountRequest
) -> RusotoFuture<CreateAccountResponse, CreateAccountError>

Creates an Amazon Chime account under the administrator's AWS account. Only Team account types are currently supported for this action. For more information about different account types, see Managing Your Amazon Chime Accounts in the Amazon Chime Administration Guide.

fn delete_account(
&self,
input: DeleteAccountRequest
) -> RusotoFuture<DeleteAccountResponse, DeleteAccountError>

Deletes the specified Amazon Chime account. You must suspend all users before deleting a Team account. You can use the BatchSuspendUser action to do so.

For EnterpriseLWA and EnterpriseAD accounts, you must release the claimed domains for your Amazon Chime account before deletion. As soon as you release the domain, all users under that account are suspended.

Deleted accounts appear in your Disabled accounts list for 90 days. To restore a deleted account from your Disabled accounts list, you must contact AWS Support.

After 90 days, deleted accounts are permanently removed from your Disabled accounts list.

fn get_account(
&self,
input: GetAccountRequest
) -> RusotoFuture<GetAccountResponse, GetAccountError>

Retrieves details for the specified Amazon Chime account, such as account type and supported licenses.

fn get_account_settings(
&self,
input: GetAccountSettingsRequest
) -> RusotoFuture<GetAccountSettingsResponse, GetAccountSettingsError>

Retrieves account settings for the specified Amazon Chime account ID, such as remote control and dial out settings. For more information about these settings, see Use the Policies Page in the Amazon Chime Administration Guide.

fn get_user(
&self,
input: GetUserRequest
) -> RusotoFuture<GetUserResponse, GetUserError>

Retrieves details for the specified user ID, such as primary email address, license type, and personal meeting PIN.

To retrieve user details with an email address instead of a user ID, use the ListUsers action, and then filter by email address.

fn invite_users(
&self,
input: InviteUsersRequest
) -> RusotoFuture<InviteUsersResponse, InviteUsersError>

Sends email invites to as many as 50 users, inviting them to the specified Amazon Chime Team account. Only Team account types are currently supported for this action.

fn list_accounts(
&self,
input: ListAccountsRequest
) -> RusotoFuture<ListAccountsResponse, ListAccountsError>

Lists the Amazon Chime accounts under the administrator's AWS account. You can filter accounts by account name prefix. To find out which Amazon Chime account a user belongs to, you can filter by the user's email address, which returns one account result.

fn list_users(
&self,
input: ListUsersRequest
) -> RusotoFuture<ListUsersResponse, ListUsersError>

Lists the users that belong to the specified Amazon Chime account. You can specify an email address to list only the user that the email address belongs to.

fn logout_user(
&self,
input: LogoutUserRequest
) -> RusotoFuture<LogoutUserResponse, LogoutUserError>

Logs out the specified user from all of the devices they are currently logged into.

fn reset_personal_pin(
&self,
input: ResetPersonalPINRequest
) -> RusotoFuture<ResetPersonalPINResponse, ResetPersonalPINError>

Resets the personal meeting PIN for the specified user on an Amazon Chime account. Returns the User object with the updated personal meeting PIN.

fn update_account(
&self,
input: UpdateAccountRequest
) -> RusotoFuture<UpdateAccountResponse, UpdateAccountError>

Updates account details for the specified Amazon Chime account. Currently, only account name updates are supported for this action.

fn update_account_settings(
&self,
input: UpdateAccountSettingsRequest
) -> RusotoFuture<UpdateAccountSettingsResponse, UpdateAccountSettingsError>

Updates the settings for the specified Amazon Chime account. You can update settings for remote control of shared screens, or for the dial-out option. For more information about these settings, see Use the Policies Page in the Amazon Chime Administration Guide.

fn update_user(
&self,
input: UpdateUserRequest
) -> RusotoFuture<UpdateUserResponse, UpdateUserError>

Updates user details for a specified user ID. Currently, only LicenseType updates are supported for this action.

Loading content...