Crate rusoto_chime

The Amazon Chime API (application programming interface) is designed for administrators to use to perform key tasks, such as creating and managing Amazon Chime accounts and users. This guide provides detailed information about the Amazon Chime API, including operations, types, inputs and outputs, and error codes.

You can use an AWS SDK, the AWS Command Line Interface (AWS CLI), or the REST API to make API calls. We recommend using an AWS SDK or the AWS CLI. Each API operation includes links to information about using it with a language-specific AWS SDK or the AWS CLI.

Using an AWS SDK

You don't need to write code to calculate a signature for request authentication. The SDK clients authenticate your requests by using access keys that you provide. For more information about AWS SDKs, see the AWS Developer Center.

Using the AWS CLI

Use your access keys with the AWS CLI to make API calls. For information about setting up the AWS CLI, see Installing the AWS Command Line Interface in the AWS Command Line Interface User Guide. For a list of available Amazon Chime commands, see the Amazon Chime commands in the AWS CLI Command Reference.

Using REST API

If you use REST to make API calls, you must authenticate your request by providing a signature. Amazon Chime supports signature version 4. For more information, see Signature Version 4 Signing Process in the Amazon Web Services General Reference.

When making REST API calls, use the service name chime and REST endpoint https://service.chime.aws.amazon.com.

Administrative permissions are controlled using AWS Identity and Access Management (IAM). For more information, see Control Access to the Amazon Chime Console in the Amazon Chime Administration Guide.

If you're using the service, you're probably looking for ChimeClient and Chime.

Structs

Account
The Amazon Chime account details. An AWS account can have multiple Amazon Chime accounts.

AccountSettings
Settings related to the Amazon Chime account. This includes settings that start or stop remote control of shared screens, or start or stop the dial-out option in the Amazon Chime web application. For more information about these settings, see Use the Policies Page in the Amazon Chime Administration Guide.

AssociatePhoneNumberWithUserRequest

AssociatePhoneNumberWithUserResponse

AssociatePhoneNumbersWithVoiceConnectorRequest

AssociatePhoneNumbersWithVoiceConnectorResponse

BatchDeletePhoneNumberRequest

BatchDeletePhoneNumberResponse

BatchSuspendUserRequest

BatchSuspendUserResponse

BatchUnsuspendUserRequest

BatchUnsuspendUserResponse

BatchUpdatePhoneNumberRequest

BatchUpdatePhoneNumberResponse

BatchUpdateUserRequest

BatchUpdateUserResponse

Bot
A resource that allows Enterprise account administrators to configure an interface to receive events from Amazon Chime.

BusinessCallingSettings
The Amazon Chime Business Calling settings for the administrator's AWS account. Includes any Amazon S3 buckets designated for storing call detail records.

ChimeClient
A client for the Amazon Chime API.

CreateAccountRequest

CreateAccountResponse

CreateBotRequest

CreateBotResponse

CreatePhoneNumberOrderRequest

CreatePhoneNumberOrderResponse

CreateVoiceConnectorRequest

CreateVoiceConnectorResponse

Credential
The SIP credentials used to authenticate requests to your Amazon Chime Voice Connector.

DeleteAccountRequest

DeleteAccountResponse

DeleteEventsConfigurationRequest

DeletePhoneNumberRequest

DeleteVoiceConnectorOriginationRequest

DeleteVoiceConnectorRequest

DeleteVoiceConnectorTerminationCredentialsRequest

DeleteVoiceConnectorTerminationRequest

DisassociatePhoneNumberFromUserRequest

DisassociatePhoneNumberFromUserResponse

DisassociatePhoneNumbersFromVoiceConnectorRequest

DisassociatePhoneNumbersFromVoiceConnectorResponse

EventsConfiguration
The configuration that allows a bot to receive outgoing events. Can be either an HTTPS endpoint or a Lambda function ARN.

GetAccountRequest

GetAccountResponse

GetAccountSettingsRequest

GetAccountSettingsResponse

GetBotRequest

GetBotResponse

GetEventsConfigurationRequest

GetEventsConfigurationResponse

GetGlobalSettingsResponse

GetPhoneNumberOrderRequest

GetPhoneNumberOrderResponse

GetPhoneNumberRequest

GetPhoneNumberResponse

GetUserRequest

GetUserResponse

GetUserSettingsRequest

GetUserSettingsResponse

GetVoiceConnectorOriginationRequest

GetVoiceConnectorOriginationResponse

GetVoiceConnectorRequest

GetVoiceConnectorResponse

GetVoiceConnectorTerminationHealthRequest

GetVoiceConnectorTerminationHealthResponse

GetVoiceConnectorTerminationRequest

GetVoiceConnectorTerminationResponse

Invite
Invitation object returned after emailing users to invite them to join the Amazon Chime Team account.

InviteUsersRequest

InviteUsersResponse

ListAccountsRequest

ListAccountsResponse

ListBotsRequest

ListBotsResponse

ListPhoneNumberOrdersRequest

ListPhoneNumberOrdersResponse

ListPhoneNumbersRequest

ListPhoneNumbersResponse

ListUsersRequest

ListUsersResponse

ListVoiceConnectorTerminationCredentialsRequest

ListVoiceConnectorTerminationCredentialsResponse

ListVoiceConnectorsRequest

ListVoiceConnectorsResponse

LogoutUserRequest

LogoutUserResponse

OrderedPhoneNumber
A phone number for which an order has been placed.

Origination
Origination settings enable your SIP hosts to receive inbound calls using your Amazon Chime Voice Connector.

OriginationRoute
Origination routes define call distribution properties for your SIP hosts to receive inbound calls using your Amazon Chime Voice Connector. Limit: 10 origination routes per Amazon Chime Voice Connector.

PhoneNumber
A phone number used for Amazon Chime Business Calling or an Amazon Chime Voice Connector.

PhoneNumberAssociation
The phone number associations, such as Amazon Chime account ID, Amazon Chime user ID, or Amazon Chime Voice Connector ID.

PhoneNumberCapabilities
The phone number capabilities, such as enabled inbound and outbound calling and text messaging.

PhoneNumberError
If the phone number action fails for one or more of the phone numbers in the request, a list of the phone numbers is returned, along with error codes and error messages.

PhoneNumberOrder
The details of a phone number order created for Amazon Chime.

PutEventsConfigurationRequest

PutEventsConfigurationResponse

PutVoiceConnectorOriginationRequest

PutVoiceConnectorOriginationResponse

PutVoiceConnectorTerminationCredentialsRequest

PutVoiceConnectorTerminationRequest

PutVoiceConnectorTerminationResponse

RegenerateSecurityTokenRequest

RegenerateSecurityTokenResponse

ResetPersonalPINRequest

ResetPersonalPINResponse

RestorePhoneNumberRequest

RestorePhoneNumberResponse

SearchAvailablePhoneNumbersRequest

SearchAvailablePhoneNumbersResponse

TelephonySettings
Settings that allow management of telephony permissions for an Amazon Chime user, such as inbound and outbound calling and text messaging.

Termination
Termination settings enable your SIP hosts to make outbound calls using your Amazon Chime Voice Connector.

TerminationHealth
The termination health details, including the source IP address and timestamp of the last successful SIP OPTIONS message from your SIP infrastructure.

UpdateAccountRequest

UpdateAccountResponse

UpdateAccountSettingsRequest

UpdateAccountSettingsResponse

UpdateBotRequest

UpdateBotResponse

UpdateGlobalSettingsRequest

UpdatePhoneNumberRequest

UpdatePhoneNumberRequestItem
The phone number ID and product type fields to update, used with the BatchUpdatePhoneNumber and UpdatePhoneNumber actions.

UpdatePhoneNumberResponse

UpdateUserRequest

UpdateUserRequestItem
The user ID and user fields to update, used with the BatchUpdateUser action.

UpdateUserResponse

UpdateUserSettingsRequest

UpdateVoiceConnectorRequest

UpdateVoiceConnectorResponse

User
The user on the Amazon Chime account.

UserError
The list of errors returned when errors are encountered during the BatchSuspendUser, BatchUnsuspendUser, or BatchUpdateUser actions. This includes user IDs, error codes, and error messages.

UserSettings
Settings associated with an Amazon Chime user, including inbound and outbound calling and text messaging.

VoiceConnector
The Amazon Chime Voice Connector configuration, including outbound host name and encryption settings.

VoiceConnectorSettings
The Amazon Chime Voice Connector settings. Includes any Amazon S3 buckets designated for storing call detail records.

Enums

AssociatePhoneNumberWithUserError
Errors returned by AssociatePhoneNumberWithUser

AssociatePhoneNumbersWithVoiceConnectorError
Errors returned by AssociatePhoneNumbersWithVoiceConnector

BatchDeletePhoneNumberError
Errors returned by BatchDeletePhoneNumber

BatchSuspendUserError
Errors returned by BatchSuspendUser

BatchUnsuspendUserError
Errors returned by BatchUnsuspendUser

BatchUpdatePhoneNumberError
Errors returned by BatchUpdatePhoneNumber

BatchUpdateUserError
Errors returned by BatchUpdateUser

CreateAccountError
Errors returned by CreateAccount

CreateBotError
Errors returned by CreateBot

CreatePhoneNumberOrderError
Errors returned by CreatePhoneNumberOrder

CreateVoiceConnectorError
Errors returned by CreateVoiceConnector

DeleteAccountError
Errors returned by DeleteAccount

DeleteEventsConfigurationError
Errors returned by DeleteEventsConfiguration

DeletePhoneNumberError
Errors returned by DeletePhoneNumber

DeleteVoiceConnectorError
Errors returned by DeleteVoiceConnector

DeleteVoiceConnectorOriginationError
Errors returned by DeleteVoiceConnectorOrigination

DeleteVoiceConnectorTerminationCredentialsError
Errors returned by DeleteVoiceConnectorTerminationCredentials

DeleteVoiceConnectorTerminationError
Errors returned by DeleteVoiceConnectorTermination

DisassociatePhoneNumberFromUserError
Errors returned by DisassociatePhoneNumberFromUser

DisassociatePhoneNumbersFromVoiceConnectorError
Errors returned by DisassociatePhoneNumbersFromVoiceConnector

GetAccountError
Errors returned by GetAccount

GetAccountSettingsError
Errors returned by GetAccountSettings

GetBotError
Errors returned by GetBot

GetEventsConfigurationError
Errors returned by GetEventsConfiguration

GetGlobalSettingsError
Errors returned by GetGlobalSettings

GetPhoneNumberError
Errors returned by GetPhoneNumber

GetPhoneNumberOrderError
Errors returned by GetPhoneNumberOrder

GetUserError
Errors returned by GetUser

GetUserSettingsError
Errors returned by GetUserSettings

GetVoiceConnectorError
Errors returned by GetVoiceConnector

GetVoiceConnectorOriginationError
Errors returned by GetVoiceConnectorOrigination

GetVoiceConnectorTerminationError
Errors returned by GetVoiceConnectorTermination

GetVoiceConnectorTerminationHealthError
Errors returned by GetVoiceConnectorTerminationHealth

InviteUsersError
Errors returned by InviteUsers

ListAccountsError
Errors returned by ListAccounts

ListBotsError
Errors returned by ListBots

ListPhoneNumberOrdersError
Errors returned by ListPhoneNumberOrders

ListPhoneNumbersError
Errors returned by ListPhoneNumbers

ListUsersError
Errors returned by ListUsers

ListVoiceConnectorTerminationCredentialsError
Errors returned by ListVoiceConnectorTerminationCredentials

ListVoiceConnectorsError
Errors returned by ListVoiceConnectors

LogoutUserError
Errors returned by LogoutUser

PutEventsConfigurationError
Errors returned by PutEventsConfiguration

PutVoiceConnectorOriginationError
Errors returned by PutVoiceConnectorOrigination

PutVoiceConnectorTerminationCredentialsError
Errors returned by PutVoiceConnectorTerminationCredentials

PutVoiceConnectorTerminationError
Errors returned by PutVoiceConnectorTermination

RegenerateSecurityTokenError
Errors returned by RegenerateSecurityToken

ResetPersonalPINError
Errors returned by ResetPersonalPIN

RestorePhoneNumberError
Errors returned by RestorePhoneNumber

SearchAvailablePhoneNumbersError
Errors returned by SearchAvailablePhoneNumbers

UpdateAccountError
Errors returned by UpdateAccount

UpdateAccountSettingsError
Errors returned by UpdateAccountSettings

UpdateBotError
Errors returned by UpdateBot

UpdateGlobalSettingsError
Errors returned by UpdateGlobalSettings

UpdatePhoneNumberError
Errors returned by UpdatePhoneNumber

UpdateUserError
Errors returned by UpdateUser

UpdateUserSettingsError
Errors returned by UpdateUserSettings

UpdateVoiceConnectorError
Errors returned by UpdateVoiceConnector

Traits

Chime
Trait representing the capabilities of the Amazon Chime API. Amazon Chime clients implement this trait.