Struct rusoto_chime::ChimeClient

pub struct ChimeClient { /* fields omitted */ }

A client for the Amazon Chime API.

Methods

impl ChimeClient

pub fn new(region: Region) -> ChimeClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> ChimeClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,