Trait rusoto_chime::Chime

pub trait Chime {
    fn associate_phone_number_with_user(
        &self, 
        input: AssociatePhoneNumberWithUserRequest
    ) -> RusotoFuture<AssociatePhoneNumberWithUserResponse, AssociatePhoneNumberWithUserError>;

    fn associate_phone_numbers_with_voice_connector(
        &self, 
        input: AssociatePhoneNumbersWithVoiceConnectorRequest
    ) -> RusotoFuture<AssociatePhoneNumbersWithVoiceConnectorResponse, AssociatePhoneNumbersWithVoiceConnectorError>;

    fn batch_delete_phone_number(
        &self, 
        input: BatchDeletePhoneNumberRequest
    ) -> RusotoFuture<BatchDeletePhoneNumberResponse, BatchDeletePhoneNumberError>;

    fn batch_suspend_user(
        &self, 
        input: BatchSuspendUserRequest
    ) -> RusotoFuture<BatchSuspendUserResponse, BatchSuspendUserError>;

    fn batch_unsuspend_user(
        &self, 
        input: BatchUnsuspendUserRequest
    ) -> RusotoFuture<BatchUnsuspendUserResponse, BatchUnsuspendUserError>;

    fn batch_update_phone_number(
        &self, 
        input: BatchUpdatePhoneNumberRequest
    ) -> RusotoFuture<BatchUpdatePhoneNumberResponse, BatchUpdatePhoneNumberError>;

    fn batch_update_user(
        &self, 
        input: BatchUpdateUserRequest
    ) -> RusotoFuture<BatchUpdateUserResponse, BatchUpdateUserError>;

    fn create_account(
        &self, 
        input: CreateAccountRequest
    ) -> RusotoFuture<CreateAccountResponse, CreateAccountError>;

    fn create_bot(
        &self, 
        input: CreateBotRequest
    ) -> RusotoFuture<CreateBotResponse, CreateBotError>;

    fn create_phone_number_order(
        &self, 
        input: CreatePhoneNumberOrderRequest
    ) -> RusotoFuture<CreatePhoneNumberOrderResponse, CreatePhoneNumberOrderError>;

    fn create_voice_connector(
        &self, 
        input: CreateVoiceConnectorRequest
    ) -> RusotoFuture<CreateVoiceConnectorResponse, CreateVoiceConnectorError>;

    fn delete_account(
        &self, 
        input: DeleteAccountRequest
    ) -> RusotoFuture<DeleteAccountResponse, DeleteAccountError>;

    fn delete_events_configuration(
        &self, 
        input: DeleteEventsConfigurationRequest
    ) -> RusotoFuture<(), DeleteEventsConfigurationError>;

    fn delete_phone_number(
        &self, 
        input: DeletePhoneNumberRequest
    ) -> RusotoFuture<(), DeletePhoneNumberError>;

    fn delete_voice_connector(
        &self, 
        input: DeleteVoiceConnectorRequest
    ) -> RusotoFuture<(), DeleteVoiceConnectorError>;

    fn delete_voice_connector_origination(
        &self, 
        input: DeleteVoiceConnectorOriginationRequest
    ) -> RusotoFuture<(), DeleteVoiceConnectorOriginationError>;

    fn delete_voice_connector_termination(
        &self, 
        input: DeleteVoiceConnectorTerminationRequest
    ) -> RusotoFuture<(), DeleteVoiceConnectorTerminationError>;

    fn delete_voice_connector_termination_credentials(
        &self, 
        input: DeleteVoiceConnectorTerminationCredentialsRequest
    ) -> RusotoFuture<(), DeleteVoiceConnectorTerminationCredentialsError>;

    fn disassociate_phone_number_from_user(
        &self, 
        input: DisassociatePhoneNumberFromUserRequest
    ) -> RusotoFuture<DisassociatePhoneNumberFromUserResponse, DisassociatePhoneNumberFromUserError>;

    fn disassociate_phone_numbers_from_voice_connector(
        &self, 
        input: DisassociatePhoneNumbersFromVoiceConnectorRequest
    ) -> RusotoFuture<DisassociatePhoneNumbersFromVoiceConnectorResponse, DisassociatePhoneNumbersFromVoiceConnectorError>;

    fn get_account(
        &self, 
        input: GetAccountRequest
    ) -> RusotoFuture<GetAccountResponse, GetAccountError>;

    fn get_account_settings(
        &self, 
        input: GetAccountSettingsRequest
    ) -> RusotoFuture<GetAccountSettingsResponse, GetAccountSettingsError>;

    fn get_bot(
        &self, 
        input: GetBotRequest
    ) -> RusotoFuture<GetBotResponse, GetBotError>;

    fn get_events_configuration(
        &self, 
        input: GetEventsConfigurationRequest
    ) -> RusotoFuture<GetEventsConfigurationResponse, GetEventsConfigurationError>;

    fn get_global_settings(
        &self
    ) -> RusotoFuture<GetGlobalSettingsResponse, GetGlobalSettingsError>;

    fn get_phone_number(
        &self, 
        input: GetPhoneNumberRequest
    ) -> RusotoFuture<GetPhoneNumberResponse, GetPhoneNumberError>;

    fn get_phone_number_order(
        &self, 
        input: GetPhoneNumberOrderRequest
    ) -> RusotoFuture<GetPhoneNumberOrderResponse, GetPhoneNumberOrderError>;

    fn get_user(
        &self, 
        input: GetUserRequest
    ) -> RusotoFuture<GetUserResponse, GetUserError>;

    fn get_user_settings(
        &self, 
        input: GetUserSettingsRequest
    ) -> RusotoFuture<GetUserSettingsResponse, GetUserSettingsError>;

    fn get_voice_connector(
        &self, 
        input: GetVoiceConnectorRequest
    ) -> RusotoFuture<GetVoiceConnectorResponse, GetVoiceConnectorError>;

    fn get_voice_connector_origination(
        &self, 
        input: GetVoiceConnectorOriginationRequest
    ) -> RusotoFuture<GetVoiceConnectorOriginationResponse, GetVoiceConnectorOriginationError>;

    fn get_voice_connector_termination(
        &self, 
        input: GetVoiceConnectorTerminationRequest
    ) -> RusotoFuture<GetVoiceConnectorTerminationResponse, GetVoiceConnectorTerminationError>;

    fn get_voice_connector_termination_health(
        &self, 
        input: GetVoiceConnectorTerminationHealthRequest
    ) -> RusotoFuture<GetVoiceConnectorTerminationHealthResponse, GetVoiceConnectorTerminationHealthError>;

    fn invite_users(
        &self, 
        input: InviteUsersRequest
    ) -> RusotoFuture<InviteUsersResponse, InviteUsersError>;

    fn list_accounts(
        &self, 
        input: ListAccountsRequest
    ) -> RusotoFuture<ListAccountsResponse, ListAccountsError>;

    fn list_bots(
        &self, 
        input: ListBotsRequest
    ) -> RusotoFuture<ListBotsResponse, ListBotsError>;

    fn list_phone_number_orders(
        &self, 
        input: ListPhoneNumberOrdersRequest
    ) -> RusotoFuture<ListPhoneNumberOrdersResponse, ListPhoneNumberOrdersError>;

    fn list_phone_numbers(
        &self, 
        input: ListPhoneNumbersRequest
    ) -> RusotoFuture<ListPhoneNumbersResponse, ListPhoneNumbersError>;

    fn list_users(
        &self, 
        input: ListUsersRequest
    ) -> RusotoFuture<ListUsersResponse, ListUsersError>;

    fn list_voice_connector_termination_credentials(
        &self, 
        input: ListVoiceConnectorTerminationCredentialsRequest
    ) -> RusotoFuture<ListVoiceConnectorTerminationCredentialsResponse, ListVoiceConnectorTerminationCredentialsError>;

    fn list_voice_connectors(
        &self, 
        input: ListVoiceConnectorsRequest
    ) -> RusotoFuture<ListVoiceConnectorsResponse, ListVoiceConnectorsError>;

    fn logout_user(
        &self, 
        input: LogoutUserRequest
    ) -> RusotoFuture<LogoutUserResponse, LogoutUserError>;

    fn put_events_configuration(
        &self, 
        input: PutEventsConfigurationRequest
    ) -> RusotoFuture<PutEventsConfigurationResponse, PutEventsConfigurationError>;

    fn put_voice_connector_origination(
        &self, 
        input: PutVoiceConnectorOriginationRequest
    ) -> RusotoFuture<PutVoiceConnectorOriginationResponse, PutVoiceConnectorOriginationError>;

    fn put_voice_connector_termination(
        &self, 
        input: PutVoiceConnectorTerminationRequest
    ) -> RusotoFuture<PutVoiceConnectorTerminationResponse, PutVoiceConnectorTerminationError>;

    fn put_voice_connector_termination_credentials(
        &self, 
        input: PutVoiceConnectorTerminationCredentialsRequest
    ) -> RusotoFuture<(), PutVoiceConnectorTerminationCredentialsError>;

    fn regenerate_security_token(
        &self, 
        input: RegenerateSecurityTokenRequest
    ) -> RusotoFuture<RegenerateSecurityTokenResponse, RegenerateSecurityTokenError>;

    fn reset_personal_pin(
        &self, 
        input: ResetPersonalPINRequest
    ) -> RusotoFuture<ResetPersonalPINResponse, ResetPersonalPINError>;

    fn restore_phone_number(
        &self, 
        input: RestorePhoneNumberRequest
    ) -> RusotoFuture<RestorePhoneNumberResponse, RestorePhoneNumberError>;

    fn search_available_phone_numbers(
        &self, 
        input: SearchAvailablePhoneNumbersRequest
    ) -> RusotoFuture<SearchAvailablePhoneNumbersResponse, SearchAvailablePhoneNumbersError>;

    fn update_account(
        &self, 
        input: UpdateAccountRequest
    ) -> RusotoFuture<UpdateAccountResponse, UpdateAccountError>;

    fn update_account_settings(
        &self, 
        input: UpdateAccountSettingsRequest
    ) -> RusotoFuture<UpdateAccountSettingsResponse, UpdateAccountSettingsError>;

    fn update_bot(
        &self, 
        input: UpdateBotRequest
    ) -> RusotoFuture<UpdateBotResponse, UpdateBotError>;

    fn update_global_settings(
        &self, 
        input: UpdateGlobalSettingsRequest
    ) -> RusotoFuture<(), UpdateGlobalSettingsError>;

    fn update_phone_number(
        &self, 
        input: UpdatePhoneNumberRequest
    ) -> RusotoFuture<UpdatePhoneNumberResponse, UpdatePhoneNumberError>;

    fn update_user(
        &self, 
        input: UpdateUserRequest
    ) -> RusotoFuture<UpdateUserResponse, UpdateUserError>;

    fn update_user_settings(
        &self, 
        input: UpdateUserSettingsRequest
    ) -> RusotoFuture<(), UpdateUserSettingsError>;

    fn update_voice_connector(
        &self, 
        input: UpdateVoiceConnectorRequest
    ) -> RusotoFuture<UpdateVoiceConnectorResponse, UpdateVoiceConnectorError>;
}

Trait representing the capabilities of the Amazon Chime API. Amazon Chime clients implement this trait.

Required methods

fn associate_phone_number_with_user(
&self,
input: AssociatePhoneNumberWithUserRequest
) -> RusotoFuture<AssociatePhoneNumberWithUserResponse, AssociatePhoneNumberWithUserError>

Associates a phone number with the specified Amazon Chime user.

fn associate_phone_numbers_with_voice_connector(
&self,
input: AssociatePhoneNumbersWithVoiceConnectorRequest
) -> RusotoFuture<AssociatePhoneNumbersWithVoiceConnectorResponse, AssociatePhoneNumbersWithVoiceConnectorError>

Associates a phone number with the specified Amazon Chime Voice Connector.

fn batch_delete_phone_number(
&self,
input: BatchDeletePhoneNumberRequest
) -> RusotoFuture<BatchDeletePhoneNumberResponse, BatchDeletePhoneNumberError>

Moves phone numbers into the Deletion queue. Phone numbers must be disassociated from any users or Amazon Chime Voice Connectors before they can be deleted.

Phone numbers remain in the Deletion queue for 7 days before they are deleted permanently.

fn batch_suspend_user(
&self,
input: BatchSuspendUserRequest
) -> RusotoFuture<BatchSuspendUserResponse, BatchSuspendUserError>

Suspends up to 50 users from a Team or EnterpriseLWA Amazon Chime account. For more information about different account types, see Managing Your Amazon Chime Accounts in the Amazon Chime Administration Guide.

Users suspended from a Team account are dissasociated from the account, but they can continue to use Amazon Chime as free users. To remove the suspension from suspended Team account users, invite them to the Team account again. You can use the InviteUsers action to do so.

Users suspended from an EnterpriseLWA account are immediately signed out of Amazon Chime and can no longer sign in. To remove the suspension from suspended EnterpriseLWA account users, use the BatchUnsuspendUser action.

To sign out users without suspending them, use the LogoutUser action.

fn batch_unsuspend_user(
&self,
input: BatchUnsuspendUserRequest
) -> RusotoFuture<BatchUnsuspendUserResponse, BatchUnsuspendUserError>

Removes the suspension from up to 50 previously suspended users for the specified Amazon Chime EnterpriseLWA account. Only users on EnterpriseLWA accounts can be unsuspended using this action. For more information about different account types, see Managing Your Amazon Chime Accounts in the Amazon Chime Administration Guide.

Previously suspended users who are unsuspended using this action are returned to Registered status. Users who are not previously suspended are ignored.

fn batch_update_phone_number(
&self,
input: BatchUpdatePhoneNumberRequest
) -> RusotoFuture<BatchUpdatePhoneNumberResponse, BatchUpdatePhoneNumberError>

Updates phone number product types. Choose from Amazon Chime Business Calling and Amazon Chime Voice Connector product types. For toll-free numbers, you can use only the Amazon Chime Voice Connector product type.

fn batch_update_user(
&self,
input: BatchUpdateUserRequest
) -> RusotoFuture<BatchUpdateUserResponse, BatchUpdateUserError>

Updates user details within the UpdateUserRequestItem object for up to 20 users for the specified Amazon Chime account. Currently, only LicenseType updates are supported for this action.

fn create_account(
&self,
input: CreateAccountRequest
) -> RusotoFuture<CreateAccountResponse, CreateAccountError>

Creates an Amazon Chime account under the administrator's AWS account. Only Team account types are currently supported for this action. For more information about different account types, see Managing Your Amazon Chime Accounts in the Amazon Chime Administration Guide.

fn create_bot(
&self,
input: CreateBotRequest
) -> RusotoFuture<CreateBotResponse, CreateBotError>

Creates a bot for an Amazon Chime Enterprise account.

fn create_phone_number_order(
&self,
input: CreatePhoneNumberOrderRequest
) -> RusotoFuture<CreatePhoneNumberOrderResponse, CreatePhoneNumberOrderError>

Creates an order for phone numbers to be provisioned. Choose from Amazon Chime Business Calling and Amazon Chime Voice Connector product types. For toll-free numbers, you can use only the Amazon Chime Voice Connector product type.

fn create_voice_connector(
&self,
input: CreateVoiceConnectorRequest
) -> RusotoFuture<CreateVoiceConnectorResponse, CreateVoiceConnectorError>

Creates an Amazon Chime Voice Connector under the administrator's AWS account. Enabling CreateVoiceConnectorRequest$RequireEncryption configures your Amazon Chime Voice Connector to use TLS transport for SIP signaling and Secure RTP (SRTP) for media. Inbound calls use TLS transport, and unencrypted outbound calls are blocked.

fn delete_account(
&self,
input: DeleteAccountRequest
) -> RusotoFuture<DeleteAccountResponse, DeleteAccountError>

Deletes the specified Amazon Chime account. You must suspend all users before deleting a Team account. You can use the BatchSuspendUser action to do so.

For EnterpriseLWA and EnterpriseAD accounts, you must release the claimed domains for your Amazon Chime account before deletion. As soon as you release the domain, all users under that account are suspended.

Deleted accounts appear in your Disabled accounts list for 90 days. To restore a deleted account from your Disabled accounts list, you must contact AWS Support.

After 90 days, deleted accounts are permanently removed from your Disabled accounts list.

fn delete_events_configuration(
&self,
input: DeleteEventsConfigurationRequest
) -> RusotoFuture<(), DeleteEventsConfigurationError>

Deletes the events configuration that allows a bot to receive outgoing events.

fn delete_phone_number(
&self,
input: DeletePhoneNumberRequest
) -> RusotoFuture<(), DeletePhoneNumberError>

Moves the specified phone number into the Deletion queue. A phone number must be disassociated from any users or Amazon Chime Voice Connectors before it can be deleted.

Deleted phone numbers remain in the Deletion queue for 7 days before they are deleted permanently.

fn delete_voice_connector(
&self,
input: DeleteVoiceConnectorRequest
) -> RusotoFuture<(), DeleteVoiceConnectorError>

Deletes the specified Amazon Chime Voice Connector. Any phone numbers assigned to the Amazon Chime Voice Connector must be unassigned from it before it can be deleted.

fn delete_voice_connector_origination(
&self,
input: DeleteVoiceConnectorOriginationRequest
) -> RusotoFuture<(), DeleteVoiceConnectorOriginationError>

Deletes the origination settings for the specified Amazon Chime Voice Connector.

fn delete_voice_connector_termination(
&self,
input: DeleteVoiceConnectorTerminationRequest
) -> RusotoFuture<(), DeleteVoiceConnectorTerminationError>

Deletes the termination settings for the specified Amazon Chime Voice Connector.

fn delete_voice_connector_termination_credentials(
&self,
input: DeleteVoiceConnectorTerminationCredentialsRequest
) -> RusotoFuture<(), DeleteVoiceConnectorTerminationCredentialsError>

Deletes the specified SIP credentials used by your equipment to authenticate during call termination.

fn disassociate_phone_number_from_user(
&self,
input: DisassociatePhoneNumberFromUserRequest
) -> RusotoFuture<DisassociatePhoneNumberFromUserResponse, DisassociatePhoneNumberFromUserError>

Disassociates the primary provisioned phone number from the specified Amazon Chime user.

fn disassociate_phone_numbers_from_voice_connector(
&self,
input: DisassociatePhoneNumbersFromVoiceConnectorRequest
) -> RusotoFuture<DisassociatePhoneNumbersFromVoiceConnectorResponse, DisassociatePhoneNumbersFromVoiceConnectorError>

Disassociates the specified phone number from the specified Amazon Chime Voice Connector.

fn get_account(
&self,
input: GetAccountRequest
) -> RusotoFuture<GetAccountResponse, GetAccountError>

Retrieves details for the specified Amazon Chime account, such as account type and supported licenses.

fn get_account_settings(
&self,
input: GetAccountSettingsRequest
) -> RusotoFuture<GetAccountSettingsResponse, GetAccountSettingsError>

Retrieves account settings for the specified Amazon Chime account ID, such as remote control and dial out settings. For more information about these settings, see Use the Policies Page in the Amazon Chime Administration Guide.

fn get_bot(
&self,
input: GetBotRequest
) -> RusotoFuture<GetBotResponse, GetBotError>

Retrieves details for the specified bot, such as bot email address, bot type, status, and display name.

fn get_events_configuration(
&self,
input: GetEventsConfigurationRequest
) -> RusotoFuture<GetEventsConfigurationResponse, GetEventsConfigurationError>

Gets details for an events configuration that allows a bot to receive outgoing events, such as an HTTPS endpoint or Lambda function ARN.

fn get_global_settings(
&self
) -> RusotoFuture<GetGlobalSettingsResponse, GetGlobalSettingsError>

Retrieves global settings for the administrator's AWS account, such as Amazon Chime Business Calling and Amazon Chime Voice Connector settings.

fn get_phone_number(
&self,
input: GetPhoneNumberRequest
) -> RusotoFuture<GetPhoneNumberResponse, GetPhoneNumberError>

Retrieves details for the specified phone number ID, such as associations, capabilities, and product type.

fn get_phone_number_order(
&self,
input: GetPhoneNumberOrderRequest
) -> RusotoFuture<GetPhoneNumberOrderResponse, GetPhoneNumberOrderError>

Retrieves details for the specified phone number order, such as order creation timestamp, phone numbers in E.164 format, product type, and order status.

fn get_user(
&self,
input: GetUserRequest
) -> RusotoFuture<GetUserResponse, GetUserError>

Retrieves details for the specified user ID, such as primary email address, license type, and personal meeting PIN.

To retrieve user details with an email address instead of a user ID, use the ListUsers action, and then filter by email address.

fn get_user_settings(
&self,
input: GetUserSettingsRequest
) -> RusotoFuture<GetUserSettingsResponse, GetUserSettingsError>

Retrieves settings for the specified user ID, such as any associated phone number settings.

fn get_voice_connector(
&self,
input: GetVoiceConnectorRequest
) -> RusotoFuture<GetVoiceConnectorResponse, GetVoiceConnectorError>

Retrieves details for the specified Amazon Chime Voice Connector, such as timestamps, name, outbound host, and encryption requirements.

fn get_voice_connector_origination(
&self,
input: GetVoiceConnectorOriginationRequest
) -> RusotoFuture<GetVoiceConnectorOriginationResponse, GetVoiceConnectorOriginationError>

Retrieves origination setting details for the specified Amazon Chime Voice Connector.

fn get_voice_connector_termination(
&self,
input: GetVoiceConnectorTerminationRequest
) -> RusotoFuture<GetVoiceConnectorTerminationResponse, GetVoiceConnectorTerminationError>

Retrieves termination setting details for the specified Amazon Chime Voice Connector.

fn get_voice_connector_termination_health(
&self,
input: GetVoiceConnectorTerminationHealthRequest
) -> RusotoFuture<GetVoiceConnectorTerminationHealthResponse, GetVoiceConnectorTerminationHealthError>

Retrieves information about the last time a SIP OPTIONS ping was received from your SIP infrastructure for the specified Amazon Chime Voice Connector.

fn invite_users(
&self,
input: InviteUsersRequest
) -> RusotoFuture<InviteUsersResponse, InviteUsersError>

Sends email invites to as many as 50 users, inviting them to the specified Amazon Chime Team account. Only Team account types are currently supported for this action.

fn list_accounts(
&self,
input: ListAccountsRequest
) -> RusotoFuture<ListAccountsResponse, ListAccountsError>

Lists the Amazon Chime accounts under the administrator's AWS account. You can filter accounts by account name prefix. To find out which Amazon Chime account a user belongs to, you can filter by the user's email address, which returns one account result.

fn list_bots(
&self,
input: ListBotsRequest
) -> RusotoFuture<ListBotsResponse, ListBotsError>

Lists the bots associated with the administrator's Amazon Chime Enterprise account ID.

fn list_phone_number_orders(
&self,
input: ListPhoneNumberOrdersRequest
) -> RusotoFuture<ListPhoneNumberOrdersResponse, ListPhoneNumberOrdersError>

Lists the phone number orders for the administrator's Amazon Chime account.

fn list_phone_numbers(
&self,
input: ListPhoneNumbersRequest
) -> RusotoFuture<ListPhoneNumbersResponse, ListPhoneNumbersError>

Lists the phone numbers for the specified Amazon Chime account, Amazon Chime user, or Amazon Chime Voice Connector.

fn list_users(
&self,
input: ListUsersRequest
) -> RusotoFuture<ListUsersResponse, ListUsersError>

Lists the users that belong to the specified Amazon Chime account. You can specify an email address to list only the user that the email address belongs to.

fn list_voice_connector_termination_credentials(
&self,
input: ListVoiceConnectorTerminationCredentialsRequest
) -> RusotoFuture<ListVoiceConnectorTerminationCredentialsResponse, ListVoiceConnectorTerminationCredentialsError>

Lists the SIP credentials for the specified Amazon Chime Voice Connector.

fn list_voice_connectors(
&self,
input: ListVoiceConnectorsRequest
) -> RusotoFuture<ListVoiceConnectorsResponse, ListVoiceConnectorsError>

Lists the Amazon Chime Voice Connectors for the administrator's AWS account.

fn logout_user(
&self,
input: LogoutUserRequest
) -> RusotoFuture<LogoutUserResponse, LogoutUserError>

Logs out the specified user from all of the devices they are currently logged into.

fn put_events_configuration(
&self,
input: PutEventsConfigurationRequest
) -> RusotoFuture<PutEventsConfigurationResponse, PutEventsConfigurationError>

Creates an events configuration that allows a bot to receive outgoing events sent by Amazon Chime. Choose either an HTTPS endpoint or a Lambda function ARN. For more information, see Bot.

fn put_voice_connector_origination(
&self,
input: PutVoiceConnectorOriginationRequest
) -> RusotoFuture<PutVoiceConnectorOriginationResponse, PutVoiceConnectorOriginationError>

Adds origination settings for the specified Amazon Chime Voice Connector.

fn put_voice_connector_termination(
&self,
input: PutVoiceConnectorTerminationRequest
) -> RusotoFuture<PutVoiceConnectorTerminationResponse, PutVoiceConnectorTerminationError>

Adds termination settings for the specified Amazon Chime Voice Connector.

fn put_voice_connector_termination_credentials(
&self,
input: PutVoiceConnectorTerminationCredentialsRequest
) -> RusotoFuture<(), PutVoiceConnectorTerminationCredentialsError>

Adds termination SIP credentials for the specified Amazon Chime Voice Connector.

fn regenerate_security_token(
&self,
input: RegenerateSecurityTokenRequest
) -> RusotoFuture<RegenerateSecurityTokenResponse, RegenerateSecurityTokenError>

Regenerates the security token for a bot.

fn reset_personal_pin(
&self,
input: ResetPersonalPINRequest
) -> RusotoFuture<ResetPersonalPINResponse, ResetPersonalPINError>

Resets the personal meeting PIN for the specified user on an Amazon Chime account. Returns the User object with the updated personal meeting PIN.

fn restore_phone_number(
&self,
input: RestorePhoneNumberRequest
) -> RusotoFuture<RestorePhoneNumberResponse, RestorePhoneNumberError>

Moves a phone number from the Deletion queue back into the phone number Inventory.

fn search_available_phone_numbers(
&self,
input: SearchAvailablePhoneNumbersRequest
) -> RusotoFuture<SearchAvailablePhoneNumbersResponse, SearchAvailablePhoneNumbersError>

Searches phone numbers that can be ordered.

fn update_account(
&self,
input: UpdateAccountRequest
) -> RusotoFuture<UpdateAccountResponse, UpdateAccountError>

Updates account details for the specified Amazon Chime account. Currently, only account name updates are supported for this action.

fn update_account_settings(
&self,
input: UpdateAccountSettingsRequest
) -> RusotoFuture<UpdateAccountSettingsResponse, UpdateAccountSettingsError>

Updates the settings for the specified Amazon Chime account. You can update settings for remote control of shared screens, or for the dial-out option. For more information about these settings, see Use the Policies Page in the Amazon Chime Administration Guide.

fn update_bot(
&self,
input: UpdateBotRequest
) -> RusotoFuture<UpdateBotResponse, UpdateBotError>

Updates the status of the specified bot, such as starting or stopping the bot from running in your Amazon Chime Enterprise account.

fn update_global_settings(
&self,
input: UpdateGlobalSettingsRequest
) -> RusotoFuture<(), UpdateGlobalSettingsError>

Updates global settings for the administrator's AWS account, such as Amazon Chime Business Calling and Amazon Chime Voice Connector settings.

fn update_phone_number(
&self,
input: UpdatePhoneNumberRequest
) -> RusotoFuture<UpdatePhoneNumberResponse, UpdatePhoneNumberError>

Updates phone number details, such as product type, for the specified phone number ID. For toll-free numbers, you can use only the Amazon Chime Voice Connector product type.

fn update_user(
&self,
input: UpdateUserRequest
) -> RusotoFuture<UpdateUserResponse, UpdateUserError>

Updates user details for a specified user ID. Currently, only LicenseType updates are supported for this action.

fn update_user_settings(
&self,
input: UpdateUserSettingsRequest
) -> RusotoFuture<(), UpdateUserSettingsError>

Updates the settings for the specified user, such as phone number settings.

fn update_voice_connector(
&self,
input: UpdateVoiceConnectorRequest
) -> RusotoFuture<UpdateVoiceConnectorResponse, UpdateVoiceConnectorError>

Updates details for the specified Amazon Chime Voice Connector.

Loading content...